+++ src/services/cashflow/cashflow_transaction_service.py
@@ -4,9 +4,9 @@ from zoneinfo import ZoneInfo
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from src.common.cashflow_types import DateType
 from src.models.accounts import Account
 from src.services.cashflow.cashflow_base import CashflowBaseService
-from src.common.cashflow_types import DateType
 from src.services.feature_flags import FeatureFlagService